more documentation for feature flags and move debug override impl to debug screen
