+++ src/components/Developer/FeatureFlags.tsx
@@ -3,10 +3,33 @@ import {
   SwitchRow,
 } from "components/SharedComponents";
 import React from "react";
-import { useFeatureFlagForDebug } from "sharedHooks/useFeatureFlag";
+import { useFeatureFlag } from "sharedHooks";
+import type { FeatureFlagSlice } from "stores/createFeatureFlagSlice";
 import { FeatureFlag } from "stores/createFeatureFlagSlice";
+import useStore from "stores/useStore";
 
 import { H1, H2 } from "./DeveloperSharedComponents";
+
+// setting feature flag overrides is "dangerous" in the sense that we don't want to accidentally
+// set them in application code so this is colocated with the only code that _should_
+const useFeatureFlagForDebug = ( featureFlagKey: FeatureFlag ) => {
+  const resolvedValue = useFeatureFlag( featureFlagKey );
+
+  const featureFlagConfig = useStore( ( state: FeatureFlagSlice ) => state.featureFlagConfig );
+  const featureFlagOverrides
+    = useStore( ( state: FeatureFlagSlice ) => state.featureFlagDebugOverrides );
+  const setOverride = useStore( ( state: FeatureFlagSlice ) => state.setFeatureFlagDebugOverride );
+
+  const rawValue = featureFlagConfig[featureFlagKey];
+  const overrideValue = featureFlagOverrides[featureFlagKey];
+  return {
+    resolvedValue,
+    rawValue,
+    overrideValue,
+    setOverride: ( enabled: boolean ) => setOverride( featureFlagKey, enabled ),
+    clearOverride: () => setOverride( featureFlagKey, null ),
+  };
+};
 
 const trimFlagName = ( featureFlagKey: FeatureFlag ) => featureFlagKey.replace( "Enabled", "" );
